Add files via upload
Here there are the scripts I wrote for the Computational Astrophysics project. The ones called 'Old_version' refer to the inital part of the project. Then we realized there was an error on the sorting of the output files and wrote other scripts to fix this error and proceed with our analysis.
Hope everything is clear.